http://f.gilt.jp/projects/3-tech/issues/5261
Added the time/date display on the left side of the screen. Note that I
have formatted the display of time to YYYY/MM/DD hh:mm. I had also made
an attempt to use CSS so that the time/day display is using font similar
to the rest of the characters used in that section of the web page. (I
simply guessed how CSS works, so I do not know if this is the correct
way of using CSS)